:truck: refactor(dto): relocate 'tokenSummary' and update related imports
