+++ src/application/dtos/auth/refreshToken.dto.ts
@@ -2,3 +2,5 @@ export interface tokenSummary {
   accessToken: string
   refreshToken: string
 }
+
+export type refreshTokenDTO = Pick<tokenSummary, 'refreshToken'>
